♻️ refactor: move URL handling logic from feed to useIntentHandler hook
